fix: EnrichmentStreamFormatTests declared shape as byte[] instead of int[]
byte[] shape = [4, 4] is a 2-byte array, but the comment says "int32 shape packed as
bytes" and Buffer.BlockCopy immediately tries to copy 8 bytes out of it -- always throws
out-of-bounds regardless of environment. int[] shape = [4, 4] is 2 int32s = 8 bytes,
matching the comment and the copy length.

+++ wow-viewer/tests/WowViewer.Core.Tests/EnrichmentStreamFormatTests.cs
@@ -135,7 +135,7 @@ public class EnrichmentStreamFormatTests
         byte[] rgb = new byte[w * h * 3];
         for (int i = 0; i < rgb.Length; i++)
             rgb[i] = (byte)(i % 256);
-        byte[] shape = [4, 4]; // int32 shape packed as bytes
+        int[] shape = [4, 4]; // int32 shape packed as bytes
         byte[] shapeBytes = new byte[8];
         Buffer.BlockCopy(shape, 0, shapeBytes, 0, 8);
